requirements.txt
python-dotenv
aiogram
sqlalchemy
psycopg2-binary